........
  r3564 | mgist | 2008-07-20 18:31:44 +0200 (Sun, 20 Jul 2008) | 3 lines
  
  - Fixed lots of compile warnings.
  - Removed deprecated GetProperty* methods.
  - Fixed a bug or two.
........

+++ V1.4/plugins/tools/quests/reward_newstate.cpp
@@ -196,7 +196,7 @@ void celClassNewStateReward::TickEveryFrame ()
     t = tag->Get (params);
 
   iCelEntity *ent;
-  for (int i = entlist->GetCount()-1; i>=0; i--)
+  for (size_t i = entlist->GetCount()-1; i>=0; i--)
   {
     ent = entlist->Get(i);
     csRef<iPcQuest> pcquest = CEL_QUERY_PROPCLASS_TAG_ENT (ent, iPcQuest, t);
