+++ V1.4/plugins/tools/quests/reward_sequence.cpp
@@ -212,7 +212,7 @@ void celClassSequenceReward::Reward (iCelParameterBlock* params)
   const char* t = tag->Get (params);
   iCelEntity *ent;
   csRef<iPcQuest> quest;
-  for (int i = entlist->GetCount()-1; i>=0; i--)
+  for (size_t i = entlist->GetCount()-1; i>=0; i--)
   {
     ent = entlist->Get(i);
     quest = CEL_QUERY_PROPCLASS_TAG_ENT (ent, iPcQuest, t);
